save changes

+++ main_service/src/main/java/ru/practicum/main_service/category/service/CategoryServiceImpl.java
@@ -4,11 +4,12 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import ru.practicum.main_service.category.repository.CategoryRepository;
 import ru.practicum.main_service.category.dto.CategoryDto;
 import ru.practicum.main_service.category.dto.NewCategoryDto;
 import ru.practicum.main_service.category.mapper.CategoryMapper;
 import ru.practicum.main_service.category.model.Category;
-import ru.practicum.main_service.category.repository.CategoryRepository;
+import ru.practicum.main_service.events.repository.EventRepository;
 import ru.practicum.main_service.exceptions.ConflictException;
 import ru.practicum.main_service.exceptions.NotFoundException;
 
@@ -22,6 +23,8 @@ public class CategoryServiceImpl implements CategoryService {
     private final CategoryRepository categoryRepository;
 
     private final CategoryMapper categoryMapper;
+
+    private final EventRepository eventRepository;
 
 
     @Override
@@ -51,6 +54,9 @@ public class CategoryServiceImpl implements CategoryService {
     public void delete(Long id) {
         categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Category not found!!!"));
+        if (eventRepository.findFirstByCategoryId(id) != null) {
+            throw new ConflictException("you cant delete a category in use");
+        }
 
         categoryRepository.deleteById(id);
     }
@@ -61,7 +67,8 @@ public class CategoryServiceImpl implements CategoryService {
         Category catToUpd = categoryRepository.findById(id)
                 .orElseThrow(() -> new NotFoundException("Category not found!!!"));
 
-        if (categoryRepository.findFirstByName(categoryDto.getName()) != null) {
+        if (categoryRepository.findFirstByName(categoryDto.getName()) != null
+                && categoryRepository.findFirstByName(categoryDto.getName()).getId() != id) {
             throw new ConflictException("Category already exist");
         }
